+++ src/edu/utexas/wrap/demand/containers/AutoODHashMatrix.java
@@ -45,7 +45,7 @@ public class AutoODHashMatrix extends ModalHashMatrix {
 	@Override
 	public void put(TravelSurveyZone origin, TravelSurveyZone destination, Float demand) {
 		map.putIfAbsent(origin, new AutoDemandHashMap(getGraph(), this));
-		get(origin).put(destination, demand.doubleValue());
+		get(origin).put(destination, demand);
 	}
 
 }
